Remove block worker heartbeat timeout by default (#878)
* Remove block worker heartbeat timeout by default

When the timeout elapses, the worker doesn't die. Instead, it continues
in a crippled state where it never tries to talk to the block master
again. The property could be useful in the future if we change the code to exit
the process on timeout. For now we are better off disabling it.

* Exit when timeout occurs

* update timeout