Add embedded journal integration tests
These are smoke tests, not comprehensive. I didn't
want to add too many tests because they take a while to run
(total of 30 seconds for the two added here). In the future
we can investigate ways to make them run faster, or we can
run our integration tests separately from the PR build so
that it's acceptable for them to run a bit longer.