Fix Kerberos authentication between workers (#1620)
Worker to worker authentication with Kerberos is broken because we used request.getCapability() != null to check the capability token and the capability is never null. According to protobuf sdk the getter will return a default instance when the value is not set. This change fixes it by using the hasCapability() check instead.